backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# OpenAI & LLM
openai>=1.12.0

# Vector Database
qdrant-client>=1.10.0  # Required for Query API with hybrid search fusion

# Sparse Embeddings (BM25 for hybrid search)
fastembed>=0.3.0

# Data Processing
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.26.0

# XML Processing
lxml>=5.0.0

# CLI Tools
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Reranking (install torch CPU: pip install torch --index-url https://download.pytorch.org/whl/cpu)
FlagEmbedding>=1.2.0

# Evaluation
ragas>=0.4.0
datasets>=2.14.0

# LangGraph Agent Framework
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
